+++ sites/partners/src/listings/PaperListingForm/index.tsx
@@ -505,6 +505,11 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
    */
   const [lotteryResultsDrawer, setLotteryResultsDrawer] = useState(false)
 
+  /**
+   * Save already-live modal
+   */
+  const [listingIsAlreadyLiveModal, setListingIsAlreadyLiveModal] = useState(false)
+
   useEffect(() => {
     if (listing?.units) {
       const tempUnits = listing.units.map((unit, i) => ({
@@ -542,8 +547,12 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
     status?: ListingStatus,
     newData?: Partial<FormListing>
   ) => {
-    if (confirm) {
-      setPublishModal(true)
+    if (confirm && status === ListingStatus.active) {
+      if (listing?.status === ListingStatus.active) {
+        setListingIsAlreadyLiveModal(true)
+      } else {
+        setPublishModal(true)
+      }
       return
     }
     let formData = { ...defaultValues, ...getValues(), ...(newData || {}) }
@@ -897,6 +906,39 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
         ]}
       >
         {t("listings.publishThisListing")}
+      </Modal>
+
+      <Modal
+        open={listingIsAlreadyLiveModal}
+        title={t("t.areYouSure")}
+        ariaDescription={t("listings.listingIsAlreadyLive")}
+        onClose={() => setListingIsAlreadyLiveModal(false)}
+        actions={[
+          <Button
+            id="saveAlreadyLiveListingButtonConfirm"
+            type="button"
+            styleType={AppearanceStyleType.success}
+            onClick={() => {
+              setListingIsAlreadyLiveModal(false)
+              triggerSubmitWithStatus(false, ListingStatus.active)
+            }}
+            dataTestId={"listingIsAlreadyLiveButton"}
+          >
+            {t("t.save")}
+          </Button>,
+          <Button
+            type="button"
+            styleType={AppearanceStyleType.secondary}
+            border={AppearanceBorderType.borderless}
+            onClick={() => {
+              setListingIsAlreadyLiveModal(false)
+            }}
+          >
+            {t("t.cancel")}
+          </Button>,
+        ]}
+      >
+        {t("listings.listingIsAlreadyLive")}
       </Modal>
     </>
   )